+++ apps/sandbox/src/components/Checkbox/Checkbox.tsx
@@ -69,15 +69,15 @@ export const Checkbox = forwardRef<HTMLButtonElement, CheckboxProps>(function Ch
       disabled={disabled}
       aria-invalid={invalid || undefined}
       onClick={activate}
-      data-ds-component="Checkbox"
-      data-ds-part="root"
+      data-juro-component="Checkbox"
+      data-juro-part="root"
       className={className}
     >
-      <div data-ds-part="box">
-        <div hidden={!(checkedValue === 'checked')} data-ds-part="tick" />
-        <div hidden={!(checkedValue === 'mixed')} data-ds-part="dash" />
+      <div data-juro-part="box">
+        <div hidden={!(checkedValue === 'checked')} data-juro-part="tick" />
+        <div hidden={!(checkedValue === 'mixed')} data-juro-part="dash" />
       </div>
-      <div data-ds-part="label">{label}</div>
+      <div data-juro-part="label">{label}</div>
       {children}
     </button>
   );