+++ apps/sandbox/src/components/TabItem/TabItem.tsx
@@ -67,12 +67,12 @@ export const TabItem = forwardRef<HTMLButtonElement, TabItemProps>(function TabI
       tabIndex={ctx.isTabStop(value) ? 0 : -1}
       aria-controls={`${ctx.baseId}-TabPanel-${value}`}
       onClick={activate}
-      data-ds-component="TabItem"
-      data-ds-part="root"
+      data-juro-component="TabItem"
+      data-juro-part="root"
       className={className}
     >
-      <div data-ds-part="label">{label}</div>
-      <div hidden={!selected} data-ds-part="indicator" />
+      <div data-juro-part="label">{label}</div>
+      <div hidden={!selected} data-juro-part="indicator" />
       {children}
     </button>
   );